Core/Player: Extend action button value to uint64 to be able to hold battle pet guids

+++ sql/base/characters_database.sql
@@ -459,7 +459,7 @@ CREATE TABLE `character_action` (
   `guid` bigint unsigned NOT NULL DEFAULT '0',
   `spec` tinyint unsigned NOT NULL DEFAULT '0',
   `button` tinyint unsigned NOT NULL DEFAULT '0',
-  `action` int unsigned NOT NULL DEFAULT '0',
+  `action` bigint unsigned NOT NULL DEFAULT '0',
   `type` tinyint unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`guid`,`spec`,`button`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
